@@ -7,13 +7,15 @@
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^26.0.23",
-    "@types/node": "^12.20.10",
-    "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.2.4",
+    "@types/node": "^17.0.45",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "classnames": "^2.3.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5",
     "typescript-eslint-parser": "^22.0.0",
     "web-vitals": "^1.1.1"
   },
@@ -45,9 +47,9 @@
     ]
   },
   "devDependencies": {
-    "@mate-academy/eslint-config-react-typescript": "^0.0.4",
-    "eslint": "^7.25.0",
-    "gh-pages": "^3.1.0",
-    "node-sass": "^5.0.0"
+    "@mate-academy/eslint-config-react-typescript": "^1.0.11",
+    "eslint": "^7.32.0",
+    "gh-pages": "^3.2.3",
+    "node-sass": "^7.0.1"
   }
 }
